overlays: add overlay_osk.h
